% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aracne.R
\name{aracne}
\alias{aracne}
\title{ARACNe-ap}
\usage{
aracne(expr = NULL, tfs = NULL, outputFolder = "./",
  pvalue = 1e-08, seed = 1, calculateThreshold = FALSE,
  consolidate = FALSE, threads = 1, nodpi = FALSE,
  nobootstrap = FALSE, nobonferroni = FALSE)
}
\arguments{
\item{expr}{the expression file.}

\item{tfs}{the file of TF list.}

\item{outputFolder}{the output folder.}

\item{pvalue}{the p-value threshold for the MI to be significant
(1E-8 usually).}

\item{seed}{the optional seed, to make the threshold mode and the
bootstrap reproducible.}

\item{calculateThreshold}{telling Java to run it in threshold mode.}

\item{consolidate}{telling java to run aracne in consolidate mode
(that is, you point it to a directory with bootstraps, and they
will be consolidated).}

\item{threads}{the number of threads (it is used only in standard
mode, i.e. bootstrap).}

\item{nodpi}{logic, if TRUE, it tells ARACNE not to run DPI.}

\item{nobootstrap}{logic, if TRUE, it tells ARACNE not to do bootstrapping.}

\item{nobonferroni}{logic, if TRUE, it removes the Bonferroni correction.}
}
\description{
ARACNe-ap
}
\examples{
\dontrun{
expr = paste0(system.file('extdata', package = 'ARACNe'), "/matrix.txt")
tfs = paste0(system.file('extdata', package = 'ARACNe'), "/tfs.txt")
# 1. calculate threshold with a fixed seed
t = aracne(expr = expr, tfs = tfs, outputFolder = "./outputFoulder/",
           calculateThreshold = T)

# 2. run ARACNe on a single bootstrap
net = aracne(expr = expr, tfs = tfs, outputFolder = "./outputFoulder/")

# 3. run 100 reproducible bootstraps
nets = lapply(setNames(1:100, paste0("b", 1:100)), function(x) {
 aracne(expr = expr, tfs = tfs,
        outputFolder = "./outputFoulder/", seed = x)})

# 4. consolidate bootstraps in the output folder
net = aracne(outputFolder = "./outputFoulder/",
             consolidate = TRUE)

# 5. run a single ARACNE with no bootstrap and no DPI (data processing
inequality)
net = aracne(expr = expr, tfs = tfs,
             outputFolder = "./outputFoulder/",
             nobootstrap = TRUE, nodpi = TRUE)

# 6. consolidate bootstraps without Bonferroni correction
net = aracne(outputFolder = "./outputFoulder/",
             consolidate = TRUE, nobonferroni = TRUE)
}
}
